+++ lib/src/widgets/karee_material_app.dart
@@ -135,9 +135,6 @@ class KareeMaterialApp extends StatelessWidget {
     }, test: (exception) => exception is TranslationFileNotExists);
 
     ErrorWidget.builder = (FlutterErrorDetails detail) {
-      print(detail.exception);
-      print(detail.summary.name);
-      print(detail.stack);
       return KareeRouterErrorWidget(
           detail.summary.name,
           detail.stack,